adding new city

+++ app/src/main/java/com/hodinv/weatherforecast/data/MainInfo.kt
@@ -5,8 +5,8 @@ package com.hodinv.weatherforecast.data
  */
 class MainInfo() {
     var temp: Float = 0.0f
-    var pressure: Int = 0
-    var humidity: Int = 0
+    var pressure: Float = 0f
+    var humidity: Float = 0f
     var temp_min: Float = 0.0f
     var temp_max: Float = 0.0f
 }
